Moved code into class services so it isn't repeating in the unregister update
